requirements.txt
fastapi
uvicorn[standard]
requests
google-cloud-vision
google-cloud-aiplatform
pydantic
Pillow
vertexai
piexif
orjson
protobuf